% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finalise_plot.R
\name{create_footer}
\alias{create_footer}
\title{Make the footer}
\usage{
create_footer(source_name, logo_image_path)
}
\arguments{
\item{source_name}{the text to be used for the source reference - may want to drop this in future}

\item{logo_image_path}{File path for the logo image you want to use in the right hand side of your chart,}
}
\description{
Running this will create the footer, including the required image/logo
}
\keyword{create_footer}
