% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finalise_plot.R
\name{figure_caption}
\alias{figure_caption}
\title{Make the figure caption}
\usage{
figure_caption(caption, figure_num = "", width = 640)
}
\arguments{
\item{caption}{the text to be used for the caption - can be left blank}

\item{figure_num}{The number, or figure reference}

\item{width}{The width of the figure in pixels - defaults to 640}
}
\description{
Running this will create the figure caption, including the figure number and appropriate formatting
}
\keyword{create_footer}
