% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finalise_plot.R
\name{finalise_plot}
\alias{finalise_plot}
\title{Arrange alignment and save TfL ggplot chart}
\usage{
finalise_plot(plot_name, caption = "", figure_number = "",
  save_filepath = file.path(Sys.getenv("TMPDIR"), "tmp-nc.png"),
  width_pixels = 640, height_pixels = 450)
}
\arguments{
\item{plot_name}{The variable name of the plot you have created that you want to format and save}

\item{caption}{The text you want to come in your figure caption (top left)}

\item{figure_number}{The figure value to associate with your figure}

\item{save_filepath}{Exact filepath that you want the plot to be saved to}

\item{width_pixels}{Width in pixels that you want to save your chart to - defaults to 640}

\item{height_pixels}{Height in pixels that you want to save your chart to - defaults to 450}
}
\value{
(Invisibly) an updated ggplot object.
}
\description{
Running this function will save your plot with the assumed guidelines from the TfL Digital Design guidelines
It will left align your title, subtitle and source, add the TfL blocks at the bottom right and save it to your specified location.
}
\keyword{finalise_plot}
