% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finalise_plot.R
\name{left_align}
\alias{left_align}
\title{Left align}
\usage{
left_align(plot_name, pieces)
}
\arguments{
\item{plot_name}{The variable name of the plot you have created that you want to format and save}

\item{pieces}{elements to be left aligned}
}
\description{
Running this will adjust the appearance of the chart
}
\keyword{left_align}
