% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finalise_plot.R
\name{save_plot}
\alias{save_plot}
\title{Export the plot}
\usage{
save_plot(plot_grid, width, height, save_filepath, bg)
}
\arguments{
\item{plot_grid}{arrange plot using a grid}

\item{width}{width of plot in pixels}

\item{height}{height of plot in pixels}

\item{save_filepath}{the filepath at which the plot will be saved}
}
\description{
Running this will export the plot using ggsave
}
\keyword{save_plot}
