% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales.R
\name{scale_color_tfl}
\alias{scale_color_tfl}
\title{Color scale constructor for tfl colours}
\usage{
scale_color_tfl(palette = "underground", discrete = TRUE,
  reverse = FALSE, ...)
}
\arguments{
\item{palette}{Character name of palette in tfl_palettes}

\item{discrete}{Boolean indicating whether colour aesthetic is discrete or not}

\item{reverse}{Boolean indicating whether the palette should be reversed}

\item{...}{Additional arguments passed to discrete_scale() or
scale_color_gradientn(), used respectively when discrete is TRUE or FALSE}
}
\description{
Color scale constructor for tfl colours
}
\keyword{scale_color_tfl}
