% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours.R
\name{tfl_cols}
\alias{tfl_cols}
\title{A collection of colours based on the colours defined in the Transport for London (TfL)
Digital Design Style Guide
http://content.tfl.gov.uk/design-style-guide.pdf
note that you must refer to http://www.tfl.gov.uk/toolkit for the latest version of the document
The available palette list is:
primary
accent
underground
modes}
\usage{
tfl_cols(...)
}
\arguments{
\item{...}{Character names of tfl_colors}
}
\description{
Function to extract TfL colors as hex codes
}
\keyword{tfl_cols}
