% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales.R
\name{tfl_pal}
\alias{tfl_pal}
\title{Return function to interpolate a tfl colour palette}
\usage{
tfl_pal(palette = "underground", reverse = FALSE, ...)
}
\arguments{
\item{palette}{Character name of palette in tfl_palettes}

\item{reverse}{Boolean indicating whether the palette should be reversed}

\item{...}{Additional arguments to pass to colorRampPalette()}
}
\description{
Return function to interpolate a tfl colour palette
}
\keyword{tfl_pal}
