% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours.R
\docType{data}
\name{tfl_palettes}
\alias{tfl_palettes}
\title{Define the tfl colour palettes as named entities}
\format{An object of class \code{list} of length 4.}
\usage{
tfl_palettes
}
\description{
Define the tfl colour palettes as named entities
}
\keyword{tfl_palettes}
