% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tfl_style.R
\name{tfl_style}
\alias{tfl_style}
\title{Add Transport for London (TfL) theme to ggplot chart}
\usage{
tfl_style()
}
\description{
This function allows you to add the TfL theme to your ggplotgraphics.
}
\keyword{tfl_style}
