% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finalise_plot.R
\name{create_footer}
\alias{create_footer}
\title{Make the footer}
\usage{
create_footer(caption, logo_image_path)
}
\arguments{
\item{caption}{the text to be used for the caption - can be left blank}

\item{logo_image_path}{File path for the logo image you want to use in the right hand side of your chart,}
}
\description{
Running this will create the footer, including the required image/logo - not currently used
}
\keyword{create_footer}
