% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours.R
\name{tfl_pal}
\alias{tfl_pal}
\title{Return function to interpolate a tfl colour palette}
\usage{
tfl_pal(palette = "underground", n, type = c("discrete", "continuous"))
}
\arguments{
\item{palette}{Character name of palette in tfl_palettes}

\item{n}{Number of desired colours}

\item{type}{Either "continous" or "discrete". Will interpolate if you use "continuous"}
}
\description{
Return function to interpolate a tfl colour palette
}
\keyword{tfl_pal}
